Consolidate root and service scripts
