+++ .gitignore
@@ -2,6 +2,8 @@ __pycache__/
 *.py[cod]
 
 map.jpg
+build/
+dist/
 
 service/__pycache__/
 service/build/
