+++ maps.py
@@ -1,420 +1,1 @@
-'''
-Module to store map metadata for lookup in the mapinfo module
-
-example maps entry: {'0707937f153ccc5d': {'name': 'Kursk',
-                             'ULHC_lat': 51.16278580067218,
-                             'ULHC_lon': 36.906235369488115,
-                             'size_km' : 65},
-                     ...}
-'''
-
-
-maps = {'0707937f153ccc5d': {'name': 'Kursk',
-                             'ULHC_lat': 51.16278580067218,
-                             'ULHC_lon': 36.906235369488115,
-                             'size_km' : 65},
-        '000000c0f0ffffff': {'name': 'Second_Battle_of_El_Alamein',
-                             'ULHC_lat': 30.462785800672183,
-                             'ULHC_lon': 27.37730048853951,
-                             'size_km' : 65},
-        '000820e066cf9fff': {'name': 'Frozen_Pass',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '00016000f3ffffb6': {'name': 'Berlin',
-                             'ULHC_lat': 52.69241915403975,
-                             'ULHC_lon': 12.939497976797764,
-                             'size_km' : 65},
-        '5f0f071008808cff': {'name': 'Berlin',
-                             'ULHC_lat': 52.69241915403975,
-                             'ULHC_lon': 12.939497976797764,
-                             'size_km' : 65},
-        'fe7e73f380070001': {'name': 'Eastern_Europe',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fffe1808081f1f3f': {'name': 'Finland',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '1f0f1f7bf3f3bbf9': {'name': 'Finland',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '8fc68480e6b6acc7': {'name': 'White_Rock_Fortress',
-                             'ULHC_lat': 52.37723024511663,
-                             'ULHC_lon': 24.130386051468452,
-                             'size_km' : 65},
-        '8080e0f4feffedff': {'name': 'Jungle',
-                             'ULHC_lat': -8.979060649367147,
-                             'ULHC_lon': 159.61888231748685,
-                             'size_km' : 65},
-        '181a72fcc5e1c103': {'name': 'Battle_of_Hurtgen_Forest',
-                             'ULHC_lat': 50.95512512075716,
-                             'ULHC_lon': 5.878575498271389,
-                             'size_km' : 65},
-        '80c48280c3dfbfff': {'name': 'Ash_River',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fcbcfef8f87cfce0': {'name': 'Karelia',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '003cf9f3f0e2d22c': {'name': 'Carpathians',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'b070e0e0c0c1cf9f': {'name': 'Kuban',
-                             'ULHC_lat': 45.04056357844995,
-                             'ULHC_lon': 38.15663550630323,
-                             'size_km' : 65},
-        '7d0d0c31b3030301': {'name': 'Mozdok',
-                             'ULHC_lat': 43.74723024511662,
-                             'ULHC_lon': 45.530795428447014,
-                             'size_km' : 65},
-        'ff0340ffffff2000': {'name': 'Normandy',
-                             'ULHC_lat': 49.63227696246796,
-                             'ULHC_lon': -1.3594883491003418,
-                             'size_km' : 65},
-        'f0f0f0a2030f1f0f': {'name': 'Fields_of_Poland',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'f0e0f0b2030f1f0f': {'name': 'Poland',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fff3f8f0c0c0f0f8': {'name': 'Port_Novorossiysk',
-                             'ULHC_lat': 45.00567346614783,
-                             'ULHC_lon': 37.350655006295746,
-                             'size_km' : 65},
-        '0cccfc7e7f7efc18': {'name': 'Advance_to_the_Rhine',
-                             'ULHC_lat': 51.20370617772062,
-                             'ULHC_lon': 6.435754396363979,
-                             'size_km' : 65},
-        '3f807efcfefcf0f1': {'name': 'Stalingrad',
-                             'ULHC_lat': 49.03954739312017,
-                             'ULHC_lon': 44.125799065790865,
-                             'size_km' : 65},
-        'fefefe00c0fefefe': {'name': 'Tunisia',
-                             'ULHC_lat': 34.38385631810749,
-                             'ULHC_lon': 9.607620383816162,
-                             'size_km' : 65},
-        'f7e54d49e8d8fc3e': {'name': 'Volokolamsk',
-                             'ULHC_lat': 56.28465547578633,
-                             'ULHC_lon': 35.53371105016213,
-                             'size_km' : 65},
-        'f8f8f8f0f0f8fcfc': {'name': 'Sinai',
-                             'ULHC_lat': 30.23374925858564,
-                             'ULHC_lon': 32.37627803962282,
-                             'size_km' : 65},
-        'f8f8f8f0f8f8fcfc': {'name': 'Sands of Sinai',
-                             'ULHC_lat': 30.23374925858564,
-                             'ULHC_lon': 32.37627803962282,
-                             'size_km' : 65},
-        '2f07030080c1e3df': {'name': '38th_Parallel',
-                             'ULHC_lat': 38.70700365064681,
-                             'ULHC_lon': 127.39023242587562,
-                             'size_km' : 65},
-        'f3f7fec341e1c000': {'name': 'Abandoned_Factory',
-                             'ULHC_lat': 57.66291098116433,
-                             'ULHC_lon': 42.45258058454457,
-                             'size_km' : 65},
-        'df60809061ffdbd0': {'name': 'Ardennes',
-                             'ULHC_lat': 50.251285800672186,
-                             'ULHC_lon': 5.237050593588114,
-                             'size_km' : 65},
-        'e0e0c0c0c0cfffff': {'name': 'Japan',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0c00476130fcfd0d': {'name': 'Japan',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '083c1e7c7870f0c0': {'name': 'Fulda',
-                             'ULHC_lat': 50.98906896351211,
-                             'ULHC_lon': 9.438324835920618,
-                             'size_km' : 65},
-        'fcfefdff7e380001': {'name': 'Middle_East',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'ef0f1f1f0f050a00': {'name': 'Maginot_Line',
-                             'ULHC_lat': 49.97091258566965,
-                             'ULHC_lon': 4.53954231640873,
-                             'size_km' : 65},
-        '00007e7e7e7e3c00': {'name': 'Campania',
-                             'ULHC_lat': 40.99554413400551,
-                             'ULHC_lon': 14.322532826882213,
-                             'size_km' : 65},
-        'fcfcfd9910befc0c': {'name': 'American_Desert',
-                             'ULHC_lat': 37.19013831774728,
-                             'ULHC_lon': -111.84834368245951,
-                             'size_km' : 65},
-        '013e3c7cfefe7c08': {'name': 'Vietnam',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0ebcfe78f87c1e0c': {'name': 'Ground_Zero',
-                             'ULHC_lat': 19.590322302910888,
-                             'ULHC_lon': 166.3231558227364,
-                             'size_km' : 65},
-        '0c3efe78f87c1e0c': {'name': 'Alaska',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '00307c7c3c3c3c00': {'name': 'Cargo_Port',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '14f4fcfcf8000001': {'name': 'Britain',
-                             'ULHC_lat': 51.41911138229066,
-                             'ULHC_lon': 0.7269576439621817,
-                             'size_km' : 65},
-        '182070381e1e0000': {'name': 'Malta',
-                             'ULHC_lat': 36.23008394553646,
-                             'ULHC_lon': 14.02448607893447,
-                             'size_km' : 65},
-        '0b0f0e1c18182000': {'name': 'Peleliu',
-                             'ULHC_lat': 7.3746069701216665,
-                             'ULHC_lon': 133.94223431827788,
-                             'size_km' : 65},
-        '080c2700c0f0ff7f': {'name': 'Guadalcanal',
-                             'ULHC_lat': -9.01556441963102,
-                             'ULHC_lon': 159.74854533169756,
-                             'size_km' : 65},
-        '07c3181c18e0e370': {'name': 'Iwo_Jima',
-                             'ULHC_lat': 25.083745648448662,
-                             'ULHC_lon': 141.02666374065078,
-                             'size_km' : 65},
-        'e7efbfb6f0f0f0e0': {'name': 'Spain',
-                             'ULHC_lat': 41.220061612629365,
-                             'ULHC_lon': 0.2981588394877504,
-                             'size_km' : 65},
-        '0781814169e7f0fb': {'name': 'Khalkhin_Gol',
-                             'ULHC_lat': 48.041953339619454,
-                             'ULHC_lon': 118.24492269637929,
-                             'size_km' : 65},
-        '0f0f0f6737030301': {'name': 'New_Guinea',
-                             'ULHC_lat': -9.134260168030261,
-                             'ULHC_lon': 146.94665506733256,
-                             'size_km' : 65},
-        '080503071f1f0f8f': {'name': 'Ruhr',
-                             'ULHC_lat': 51.73829303094487,
-                             'ULHC_lon': 6.437537416826182,
-                             'size_km' : 65},
-        'fcf8f0e0fcfcfcfe': {'name': 'Sicily',
-                             'ULHC_lat': 37.62582130005829,
-                             'ULHC_lon': 14.575726231320882,
-                             'size_km' : 65},
-        '00c00cb88cbcfcff': {'name': 'Korea',
-                             'ULHC_lat': 38.310264019718346,
-                             'ULHC_lon': 127.2747578040516,
-                             'size_km' : 65},
-        '70f8f8fe6f6f0200': {'name': 'Honolulu',
-                             'ULHC_lat': 21.742180886911395,
-                             'ULHC_lon': -158.2681626706867,
-                             'size_km' : 65},
-        '000018dc3c3fff7b': {'name': 'Afghanistan',
-                             'ULHC_lat': 35.22242668541464,
-                             'ULHC_lon': 68.95616436599424,
-                             'size_km' : 65},
-        '1f1f1c09117fe343': {'name': 'Smolensk',
-                             'ULHC_lat': 55.08537566230713,
-                             'ULHC_lon': 31.556597199003146,
-                             'size_km' : 65},
-        'fcfcd8e46c0e0080': {'name': 'African_Canyon',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'ffc300000000e3ff': {'name': 'Gorge',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '000000010383cfdf': {'name': 'Foothills',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fffc20c0d0c080e0': {'name': 'Cliffed_Coast',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '3f9f47030363c7c4': {'name': 'Alpine_Meadows',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fffff9f8d8c0c000': {'name': 'Wake_Island',
-                             'ULHC_lat': 19.590322302910888,
-                             'ULHC_lon': 166.3231558227364,
-                             'size_km' : 65},
-        'f7fef9f8d8c08000': {'name': 'Wake_Island',
-                             'ULHC_lat': 19.590322302910888,
-                             'ULHC_lon': 166.3231558227364,
-                             'size_km' : 65},
-        'f7fff9f8d8c08000': {'name': 'Wake_Island',
-                             'ULHC_lat': 19.590322302910888,
-                             'ULHC_lon': 166.3231558227364,
-                             'size_km' : 65},
-        'fe7e73f701070301': {'name': 'European_Province',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '80c0f0f8d8fcf8f0': {'name': 'Fire_Arc',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'c03c642113537176': {'name': 'Fields_of_Normandy',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '003c7e7e7e7c3c00': {'name': 'Sweden',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '9f1b3efebe660090': {'name': 'Seversk-13',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'df81d9e0e0e981db': {'name': 'Red Desert',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '98883e3e3e0e0e8c': {'name': 'Breslau',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '79c4dcf1e187e2e0': {'name': 'Spaceport',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fbffffc6c2c2c880': {'name': 'Aral_Sea',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'f0f8f8e0c0e0f8f8': {'name': 'Sun_City',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'c0f0fcfece604000': {'name': 'Abandoned_Town',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'f3f7a282d3e3e3f9': {'name': 'Pradesh',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'e4fcfcfedf1c2aef': {'name': 'Artic_Pier',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'e4fcfcfedf1c2aef': {'name': 'Artic_Polar_Base',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0f47c38008b88fe6': {'name': 'Golden_Quarry',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0f0f1fdde4e4f0f0': {'name': 'Poland_Winter',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0f0f1fdde4e4f0f0': {'name': 'Fields_of_Poland_Winter',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0f471b1f1f1f3f02': {'name': 'Maginot_Line_Winter',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0f471b1f1f1f3f02': {'name': 'Maginot_Line_Winter',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '3f1ffcc1811f23f3': {'name': 'Seversk_13_Winter',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'ffffff7f3f1f0000': {'name': 'Iberian_Castle',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fee8c39b9b828000': {'name': 'Test_Site_2271',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fcfe5e1e1c0802c3': {'name': 'Flanders',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'cf9b133330387960': {'name': 'Kamchatka',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'e1432f3c78200f7f': {'name': 'Southeastern_City',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0300321707fffe40': {'name': 'Rocky_Pillars',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'decfc3c3a1f0fc7c': {'name': 'Rocky_Canyon',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'ff03f2fe1f00807f': {'name': 'Pyrenees',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'ff03f2fe1f00807f': {'name': 'Pyrenees',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'c3c3c3c3ffffffff': {'name': 'Ladoga_Winter_1941',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'e8e020d0f880e07e': {'name': 'Moscow',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '7c7eff1f1e1c0800': {'name': 'Battle_for_Vietnam',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '07070f0f1f1f1f2f': {'name': 'Golan_Heights',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '00183c3c3c180000': {'name': 'Bourbon_Island',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '00031371fbf0e8f0': {'name': 'Laizhou_Bay',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '079b8683c5c5c2c2': {'name': 'City',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '1f1f1c09113fe303': {'name': 'Smolensk',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'fcf8f0e0fcfefcfe': {'name': 'Husky',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        'e3c188a327f399bf': {'name': 'Korsun',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-        '0787c7e3c3c0cbcf': {'name': 'Battle_for_the_Rhine',
-                             'ULHC_lat': 0.0,
-                             'ULHC_lon': 0.0,
-                             'size_km' : 65},
-}
+from warthunder_rpc.maps import *
