+++ requirements.txt
@@ -1,0 +1,8 @@
+pypresence==4.2.1
+Pillow==10.0.0
+requests==2.31.0
+pywin32==306
+psutil==7.0.0
+ImageHash==4.3.1
+simplejson==3.20.1
+pyinstaller==6.11.0
