--- service/requirements.txt
@@ -1,5 +1,0 @@
-pypresence==4.2.1
-Pillow==10.0.0
-requests==2.31.0
-pywin32==306
-nest-asyncio==1.6.0
